// types/Types.ts
export interface SingleTagType {
  _id: string;
  clerkUserId: string;
  name: string;
}